adaptation de la classe calculatorController et suppression de calculator.java

+++ src/main/java/fr/devavance/calculatrice/Addition.java
@@ -11,7 +11,7 @@ package fr.devavance.calculatrice;
 public class Addition implements Operation{
     @Override
     public Double calculer(int operande_1, int operande_2){
-        return operande_1 + operande_2;
+        return (double) operande_1 + operande_2;
     }
     @Override
     public String getNom(){
